STNG-165 Run one scenario fully end-to-end in AWS dev environment
* Refactored quite a bit, to increase on code reuse
