+++ spring-boot/src/test/java/org/dcsa/conformance/manual/ManualScenarioTest.java
@@ -6,11 +6,14 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.List;
 import java.util.stream.Stream;
 import lombok.extern.slf4j.Slf4j;
+import org.dcsa.conformance.springboot.ConformanceApplication;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.springframework.boot.test.context.SpringBootTest;
 
 @Slf4j
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT, classes = ConformanceApplication.class)
 class ManualScenarioTest extends ManualTestBase {
 
   @SuppressWarnings("unused")
